Refactor and add community exploration step
Instead of handling state somehow, we just look whether the user has
communities or not and show the username selection and community
exploration screens based on that!
