+++ mobile/views/UserOnboarding/ExploreCommunities.js
@@ -1,0 +1,16 @@
+// @flow
+import React from 'react';
+import { SafeAreaView } from 'react-native';
+import Text from '../../components/Text';
+
+class ExploreCommunities extends React.Component<{}> {
+  render() {
+    return (
+      <SafeAreaView>
+        <Text>This is some sweet community exploring</Text>
+      </SafeAreaView>
+    );
+  }
+}
+
+export default ExploreCommunities;
